Setted on the AppLockManager.swift the code related with the Biometrical from the old PasscodeManager.swift
